TDI-35893 Move getName() getPngImagePath() Constructor(String componentName) to AbstractComponentDefinition class
