+++ components-salesforce/src/main/java/org/talend/components/salesforce/SalesforceDefinition.java
@@ -13,40 +13,17 @@
 package org.talend.components.salesforce;
 
 import org.talend.components.api.component.AbstractComponentDefinition;
-import org.talend.components.api.component.ComponentImageType;
 import org.talend.components.api.properties.ComponentProperties;
 
 public abstract class SalesforceDefinition extends AbstractComponentDefinition {
-
-    protected String componentName;
-
+    
     public SalesforceDefinition(String componentName) {
-        this.componentName = componentName;
+        super(componentName);
     }
 
     @Override
     public String[] getFamilies() {
         return new String[] { "Business/Salesforce", "Cloud/Salesforce" };
-    }
-
-    @Override
-    public String getName() {
-        return componentName;
-    }
-
-    @Override
-    public String getPngImagePath(ComponentImageType imageType) {
-        switch (imageType) {
-        case PALLETE_ICON_32X32:
-            return componentName + "_icon32.png"; //$NON-NLS-1$
-        }
-        return null;
-    }
-
-    @Override
-    public String getDisplayName() {
-        // FIXME
-        return getName();
     }
 
     @SuppressWarnings("unchecked")
